fix(runtime,core): harden bounds checks, fix ODR violations, and add quiescence obligation gate
Multiple correctness and safety fixes across the runtime and core:

- asx_ids.h: add lower-bound check in asx_cancel_severity to prevent
  negative array indexing on invalid cancel kinds
- mpsc.c: add capacity check in asx_send_permit_send to prevent buffer
  overflow when queue is full (return ASX_E_CHANNEL_FULL)
- adaptive.c: clamp state_count to ASX_ADAPTIVE_MAX_ACTIONS in
  compute_expected_loss to prevent out-of-bounds reads
- adapter.c: remove duplicate asx_adapter_mode_str definition (now
  lives in vertical_adapter.c) to fix ODR linker error
- automotive_instrument.c: save previous checkpoint time before updating
  watchdog, fixing watchdog violation interval calculation (was always 0)
- hft_instrument.c: cast histogram percentile multiplication to uint64_t
  to prevent uint32 overflow on large totals
- hooks.c: remove duplicated cross-codec semantic equivalence functions
  (now in equivalence.c) to fix ODR linker violation
- quiescence.c: add obligation resolution check (Q3 gate) — regions
  with unresolved RESERVED obligations now correctly block quiescence

+++ src/runtime/hft_instrument.c
@@ -96,7 +96,7 @@ uint64_t asx_hft_histogram_percentile(const asx_hft_histogram *h,
     if (!h || h->total == 0) return 0;
     if (pct > 100) pct = 100;
 
-    target = (h->total * pct) / 100u;
+    target = (uint32_t)((uint64_t)h->total * pct / 100u);
     if (target == 0 && pct > 0) target = 1;
 
     cumulative = 0;
